+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpIsoProtocolTest.kt
@@ -1,6 +1,7 @@
 package at.asitplus.wallet.lib.openid
 
 import at.asitplus.openid.OpenIdConstants
+import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.testballoon.invoke
 import at.asitplus.testballoon.withFixtureGenerator
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithSelfSignedCert
@@ -12,10 +13,12 @@ import at.asitplus.wallet.lib.agent.toStoreCredentialInput
 import at.asitplus.wallet.lib.data.ConstantIndex.AtomicAttribute2023
 import at.asitplus.wallet.lib.data.ConstantIndex.AtomicAttribute2023.CLAIM_GIVEN_NAME
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.ISO_MDOC
-import at.asitplus.wallet.lib.data.IsoDocumentParsed
 import at.asitplus.wallet.lib.data.rfc3986.toUri
 import at.asitplus.wallet.lib.oidvci.formUrlEncode
-import at.asitplus.wallet.mdl.MobileDrivingLicenceDataElements
+import at.asitplus.wallet.lib.openid.AuthnResponseResult.SuccessIso
+import at.asitplus.wallet.lib.openid.AuthnResponseResult.VerifiableDCQLPresentationValidationResults
+import at.asitplus.wallet.lib.openid.OpenId4VpVerifier.CreationOptions.Query
+import at.asitplus.wallet.mdl.MobileDrivingLicenceDataElements.FAMILY_NAME
 import at.asitplus.wallet.mdl.MobileDrivingLicenceDataElements.GIVEN_NAME
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
 import com.benasher44.uuid.uuid4
@@ -24,8 +27,14 @@ import io.kotest.engine.runBlocking
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldBeSingleton
 import io.kotest.matchers.collections.shouldHaveSingleElement
+import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.collections.shouldNotBeEmpty
+import io.kotest.matchers.maps.shouldHaveSize
+import io.kotest.matchers.nulls.shouldNotBeNull
+import io.kotest.matchers.string.shouldNotBeEmpty
 import io.kotest.matchers.types.shouldBeInstanceOf
+import kotlinx.serialization.json.JsonArray
+import kotlinx.serialization.json.JsonObject
 
 val OpenId4VpIsoProtocolTest by testSuite {
 
@@ -78,57 +87,65 @@ val OpenId4VpIsoProtocolTest by testSuite {
         }
     } - {
         "test with Fragment for mDL" {
-            runProcess(
-                it.verifierOid4vp,
-                it.walletUrl,
-                RequestOptions(
-                    credentials = setOf(
-                        RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(GIVEN_NAME))
-                    )
-                ),
-                it.holderOid4vp
-            ).apply {
-                validItems.shouldNotBeEmpty()
-                invalidItems.shouldBeEmpty()
-            }
+            val requestOptions = RequestOptions(
+                credentials = setOf(
+                    RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(GIVEN_NAME))
+                )
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
+            val authnResponse = it.holderOid4vp
+                .createAuthnResponse(authnRequest).getOrThrow()
+                .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
+            it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
+                .shouldBeInstanceOf<SuccessIso>()
+                .documents.first().apply {
+                    validItems.shouldNotBeEmpty()
+                    invalidItems.shouldBeEmpty()
+                }
         }
 
         "test with Fragment for custom attributes" {
-            runProcess(
-                it.verifierOid4vp,
-                it.walletUrl,
-                RequestOptions(
-                    credentials = setOf(
-                        RequestOptionsCredential(AtomicAttribute2023, ISO_MDOC, setOf(CLAIM_GIVEN_NAME))
-                    )
-                ),
-                it.holderOid4vp
-            ).apply {
-                validItems.shouldNotBeEmpty()
-                invalidItems.shouldBeEmpty()
-            }
+            val requestOptions = RequestOptions(
+                credentials = setOf(
+                    RequestOptionsCredential(AtomicAttribute2023, ISO_MDOC, setOf(CLAIM_GIVEN_NAME))
+                )
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
+            val authnResponse = it.holderOid4vp
+                .createAuthnResponse(authnRequest).getOrThrow()
+                .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
+            it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
+                .shouldBeInstanceOf<SuccessIso>()
+                .documents.first().apply {
+                    validItems.shouldNotBeEmpty()
+                    invalidItems.shouldBeEmpty()
+                }
         }
 
         "Selective Disclosure with mDL" {
-            val requestedClaim = MobileDrivingLicenceDataElements.FAMILY_NAME
-            runProcess(
-                it.verifierOid4vp,
-                it.walletUrl,
-                RequestOptions(
-                    credentials = setOf(
-                        RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(requestedClaim))
-                    )
-                ),
-                it.holderOid4vp,
-            ).apply {
-                validItems.shouldBeSingleton()
-                validItems.shouldHaveSingleElement { it.elementIdentifier == requestedClaim }
-                invalidItems.shouldBeEmpty()
-            }
+            val requestedClaim = FAMILY_NAME
+            val requestOptions = RequestOptions(
+                credentials = setOf(
+                    RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(requestedClaim))
+                )
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
+            val authnResponse = it.holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
+                .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
+            it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
+                .shouldBeInstanceOf<SuccessIso>()
+                .documents.first().apply {
+                    validItems.shouldBeSingleton()
+                    validItems.shouldHaveSingleElement { it.elementIdentifier == requestedClaim }
+                    invalidItems.shouldBeEmpty()
+                }
         }
 
         "Selective Disclosure with mDL (ISO/IEC 18013-7:2024 Annex B)" {
-            val requestedClaim = MobileDrivingLicenceDataElements.FAMILY_NAME
+            val requestedClaim = FAMILY_NAME
             val requestOptions = RequestOptions(
                 credentials = setOf(
                     RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(requestedClaim))
@@ -136,9 +153,8 @@ val OpenId4VpIsoProtocolTest by testSuite {
                 responseMode = OpenIdConstants.ResponseMode.DirectPost,
                 responseUrl = "https://example.com/response",
             )
-            val authnRequest = it.verifierOid4vp.createAuthnRequest(
-                requestOptions, OpenId4VpVerifier.CreationOptions.Query(it.walletUrl)
-            ).getOrThrow().url
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
             //println("this is the request:\n$authnRequest")
 
             val authnResponse = it.holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
@@ -148,7 +164,7 @@ val OpenId4VpIsoProtocolTest by testSuite {
             //println("this is the response:\n$input")
 
             it.verifierOid4vp.validateAuthnResponse(input)
-                .shouldBeInstanceOf<AuthnResponseResult.SuccessIso>()
+                .shouldBeInstanceOf<SuccessIso>()
                 .documents.first().apply {
                     validItems.shouldBeSingleton()
                     validItems.shouldHaveSingleElement { it.elementIdentifier == requestedClaim }
@@ -157,7 +173,7 @@ val OpenId4VpIsoProtocolTest by testSuite {
         }
 
         "Selective Disclosure with mDL and encryption (ISO/IEC 18013-7:2024 Annex B)" {
-            val requestedClaim = MobileDrivingLicenceDataElements.FAMILY_NAME
+            val requestedClaim = FAMILY_NAME
             val requestOptions = RequestOptions(
                 credentials = setOf(
                     RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(requestedClaim))
@@ -167,7 +183,7 @@ val OpenId4VpIsoProtocolTest by testSuite {
                 encryption = true
             )
             val authnRequest = it.verifierOid4vp.createAuthnRequest(
-                requestOptions, OpenId4VpVerifier.CreationOptions.Query(it.walletUrl)
+                requestOptions, Query(it.walletUrl)
             ).getOrThrow().url
             //println("this is the request:\n$authnRequest")
 
@@ -178,7 +194,7 @@ val OpenId4VpIsoProtocolTest by testSuite {
             //println("this is the response:\n$input")
 
             it.verifierOid4vp.validateAuthnResponse(input)
-                .shouldBeInstanceOf<AuthnResponseResult.SuccessIso>()
+                .shouldBeInstanceOf<SuccessIso>()
                 .documents.first().apply {
                     validItems.shouldBeSingleton()
                     validItems.shouldHaveSingleElement { it.elementIdentifier == requestedClaim }
@@ -186,28 +202,34 @@ val OpenId4VpIsoProtocolTest by testSuite {
                 }
         }
 
-        "Selective Disclosure with two documents and encryption (ISO/IEC 18013-7:2024 Annex B)" {
-            val mdlFamilyName = MobileDrivingLicenceDataElements.FAMILY_NAME
+        "Selective Disclosure with two documents in presentation exchange" {
+            val mdlFamilyName = FAMILY_NAME
             val atomicGivenName = CLAIM_GIVEN_NAME
             val requestOptions = RequestOptions(
                 credentials = setOf(
                     RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(mdlFamilyName)),
                     RequestOptionsCredential(AtomicAttribute2023, ISO_MDOC, setOf(atomicGivenName))
                 ),
-                responseMode = OpenIdConstants.ResponseMode.DirectPostJwt,
+                responseMode = OpenIdConstants.ResponseMode.DirectPost,
                 responseUrl = "https://example.com/response",
-                encryption = true
+                presentationMechanism = PresentationMechanismEnum.PresentationExchange
             )
-            val authnRequest = it.verifierOid4vp.createAuthnRequest(
-                requestOptions, OpenId4VpVerifier.CreationOptions.Query(it.walletUrl)
-            ).getOrThrow().url
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
 
             val authnResponse = it.holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
-                .shouldBeInstanceOf<AuthenticationResponseResult.Post>()
+                .shouldBeInstanceOf<AuthenticationResponseResult.Post>().apply {
+                    // make sure there are two device responses for two credentials returned in the presentation
+                    params["vp_token"].shouldNotBeEmpty().shouldNotBeNull().apply {
+                        joseCompliantSerializer.decodeFromString<JsonArray>(this).apply {
+                            shouldHaveSize(2)
+                        }
+                    }
+                }
 
             it.verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
                 .shouldBeInstanceOf<AuthnResponseResult.VerifiablePresentationValidationResults>()
-                .validationResults.flatMap { it.shouldBeInstanceOf<AuthnResponseResult.SuccessIso>().documents }.apply {
+                .validationResults.flatMap { it.shouldBeInstanceOf<SuccessIso>().documents }.apply {
                     first { it.mso.docType == AtomicAttribute2023.isoDocType }
                         .validItems.shouldHaveSingleElement { it.elementIdentifier == atomicGivenName }
                     first { it.mso.docType == MobileDrivingLicenceScheme.isoDocType }
@@ -215,43 +237,65 @@ val OpenId4VpIsoProtocolTest by testSuite {
                 }
         }
 
-        "Selective Disclosure with mDL JSON Path syntax" {
-            runProcess(
-                it.verifierOid4vp,
-                it.walletUrl,
-                RequestOptions(
-                    credentials = setOf(
-                        RequestOptionsCredential(
-                            MobileDrivingLicenceScheme,
-                            ISO_MDOC,
-                            setOf(MobileDrivingLicenceDataElements.FAMILY_NAME)
-                        )
-                    )
+        "Selective Disclosure with two documents in DCQL" {
+            val mdlFamilyName = FAMILY_NAME
+            val atomicGivenName = CLAIM_GIVEN_NAME
+            val requestOptions = RequestOptions(
+                credentials = setOf(
+                    RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(mdlFamilyName)),
+                    RequestOptionsCredential(AtomicAttribute2023, ISO_MDOC, setOf(atomicGivenName))
                 ),
-                it.holderOid4vp,
-            ).apply {
-                validItems.shouldBeSingleton()
-                validItems.shouldHaveSingleElement { it.elementIdentifier == MobileDrivingLicenceDataElements.FAMILY_NAME }
-                invalidItems.shouldBeEmpty()
-            }
+                responseMode = OpenIdConstants.ResponseMode.DirectPost,
+                responseUrl = "https://example.com/response",
+                presentationMechanism = PresentationMechanismEnum.DCQL,
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
+
+            val authnResponse = it.holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
+                .shouldBeInstanceOf<AuthenticationResponseResult.Post>().apply {
+                    // make sure there are two device responses for two credentials returned in the presentation
+                    params["vp_token"].shouldNotBeEmpty().shouldNotBeNull().apply {
+                        joseCompliantSerializer.decodeFromString<JsonObject>(this).apply {
+                            shouldHaveSize(2)
+                        }
+                    }
+                }
+
+            it.verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
+                .shouldBeInstanceOf<VerifiableDCQLPresentationValidationResults>()
+                .validationResults.shouldHaveSize(2).apply {
+                    values.first { it.hasDocType(AtomicAttribute2023.isoDocType) }
+                        .shouldBeInstanceOf<SuccessIso>().documents.first()
+                        .validItems.shouldHaveSingleElement { it.elementIdentifier == atomicGivenName }
+                    values.first { it.hasDocType(MobileDrivingLicenceScheme.isoDocType) }
+                        .shouldBeInstanceOf<SuccessIso>().documents.first()
+                        .validItems.shouldHaveSingleElement { it.elementIdentifier == mdlFamilyName }
+                }
+        }
+
+        "Selective Disclosure with mDL JSON Path syntax" {
+            val requestOptions = RequestOptions(
+                credentials = setOf(
+                    RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(FAMILY_NAME))
+                )
+            )
+            val authnRequest = it.verifierOid4vp.createAuthnRequest(requestOptions, Query(it.walletUrl))
+                .getOrThrow().url
+            val authnResponse = it.holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
+                .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
+            it.verifierOid4vp.validateAuthnResponse(authnResponse.url)
+                .shouldBeInstanceOf<SuccessIso>()
+                .documents.first().apply {
+                    validItems.shouldBeSingleton()
+                    validItems.shouldHaveSingleElement { it.elementIdentifier == FAMILY_NAME }
+                    invalidItems.shouldBeEmpty()
+                }
         }
     }
 }
 
-private suspend fun runProcess(
-    verifierOid4vp: OpenId4VpVerifier,
-    walletUrl: String,
-    requestOptions: RequestOptions,
-    holderOid4vp: OpenId4VpHolder,
-): IsoDocumentParsed {
-    val authnRequest = verifierOid4vp.createAuthnRequest(
-        requestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
-    ).getOrThrow().url
+private fun AuthnResponseResult.hasDocType(docType: String): Boolean =
+    this.shouldBeInstanceOf<SuccessIso>().documents
+        .shouldBeSingleton().first().mso.docType == docType
 
-    val authnResponse = holderOid4vp.createAuthnResponse(authnRequest).getOrThrow()
-        .shouldBeInstanceOf<AuthenticationResponseResult.Redirect>()
-
-    return verifierOid4vp.validateAuthnResponse(authnResponse.url)
-        .shouldBeInstanceOf<AuthnResponseResult.SuccessIso>()
-        .documents.first()
-}