+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/HolderAgent.kt
@@ -196,35 +196,20 @@ class HolderAgent(
             )
         }.toList()
 
-        // Presentation will be one single ISO mDoc DeviceResponse, containing multiple documents
-        val isSingleIsoMdocPresentation = submissionList.all { it.second.credential is StoreEntry.Iso }
         val presentationSubmission = PresentationSubmission.fromMatches(
             presentationId = presentationDefinition.id,
-            matches = submissionList,
-            isSingleIsoMdocPresentation = isSingleIsoMdocPresentation
+            matches = submissionList
         )
-
-        val verifiablePresentations = if (isSingleIsoMdocPresentation) {
-            listOf(
-                verifiablePresentationFactory.createVerifiablePresentationForIsoCredentials(
-                    request = request,
-                    credentialAndDisclosedAttributes = submissionList
-                        .associate { it.second.credential as StoreEntry.Iso to it.second.disclosedAttributes },
-                ).getOrThrow()
-            )
-        } else {
-            submissionList.map { match ->
-                verifiablePresentationFactory.createVerifiablePresentation(
-                    request = request,
-                    credential = match.second.credential,
-                    disclosedAttributes = match.second.disclosedAttributes,
-                ).getOrThrow()
-            }
-        }
 
         PresentationResponseParameters.PresentationExchangeParameters(
             presentationSubmission = presentationSubmission,
-            presentationResults = verifiablePresentations,
+            presentationResults = submissionList.map { match ->
+                    verifiablePresentationFactory.createVerifiablePresentation(
+                        request = request,
+                        credential = match.second.credential,
+                        disclosedAttributes = match.second.disclosedAttributes,
+                    ).getOrThrow()
+                },
         )
     }
 
@@ -330,7 +315,6 @@ class HolderAgent(
     private fun PresentationSubmission.Companion.fromMatches(
         presentationId: String?,
         matches: List<Pair<String, PresentationExchangeCredentialDisclosure>>,
-        isSingleIsoMdocPresentation: Boolean,
     ) = PresentationSubmission(
         id = uuid4().toString(),
         definitionId = presentationId,
@@ -338,7 +322,7 @@ class HolderAgent(
             PresentationSubmissionDescriptor.fromMatch(
                 inputDescriptorId = match.first,
                 credential = match.second.credential,
-                index = if (matches.size == 1 || isSingleIsoMdocPresentation) null else index,
+                index = if (matches.size == 1) null else index,
             )
         },
     )